refactor(idea_service): extract standing-questions helpers (#163)
  before: 2191 lines
  after:  2160 lines (-31)

321 tests pass.
